chore: move primitives to /internal/components
