+++ src/internal/components/Dialog.tsx
