+++ src/internal/components/FocusTrap.tsx
