+++ src/swap/components/SwapSettings.tsx
@@ -1,6 +1,6 @@
-import { DismissableLayer } from '@/internal/components/primitives/DismissableLayer';
-import { FocusTrap } from '@/internal/components/primitives/FocusTrap';
-import { Popover } from '@/internal/components/primitives/Popover';
+import { DismissableLayer } from '@/internal/components/DismissableLayer';
+import { FocusTrap } from '@/internal/components/FocusTrap';
+import { Popover } from '@/internal/components/Popover';
 import { useBreakpoints } from '@/internal/hooks/useBreakpoints';
 import { useIcon } from '@/internal/hooks/useIcon';
 import { background, border, cn, pressable, text } from '@/styles/theme';
